rename to ICRProjectContract
