+++ polygon-digital-carbon/src/templates/ICRCarbonContractRegistry.ts
@@ -1,4 +1,4 @@
-import { ICRProjectToken } from '../../generated/templates'
+import { ICRProjectContract } from '../../generated/templates'
 import { ProjectCreated } from '../../generated/ICRCarbonContractRegistry/ICRCarbonContractRegistry'
 import { log } from '@graphprotocol/graph-ts'
 
@@ -6,7 +6,7 @@ export function handleNewICC(event: ProjectCreated): void {
   // Start indexing the C3T tokens; `event.params.tokenAddress` is the
   // address of the new token contract
 
-  ICRProjectToken.create(event.params.projectAddress)
+  ICRProjectContract.create(event.params.projectAddress)
 
   log.info('Created new ICR Project Datasource', [])
 }
